# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for UniSphere university social platform",
        "title": "UniSphere API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://www.unisphere.com/support",
            "email": "support@unisphere.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/pastexams/{id}": {
            "get": {
                "description": "Retrieves a specific past exam by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "past-exams"
                ],
                "summary": "Get past exam by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Past exam ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Past exam retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.PastExamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid past exam ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Past exam not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing past exam with the provided information",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "past-exams"
                ],
                "summary": "Update a past exam",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Past exam ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Year of the exam",
                        "name": "year",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Term of the exam (FALL, SPRING)",
                        "name": "term",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Department ID",
                        "name": "departmentId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Course code",
                        "name": "courseCode",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Exam title",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Instructor ID",
                        "name": "instructorId",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Exam file",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Past exam updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.PastExamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Past exam not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an existing past exam by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "past-exams"
                ],
                "summary": "Delete a past exam",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Past exam ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Past exam deleted successfully"
                    },
                    "400": {
                        "description": "Invalid past exam ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Past exam not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/past-exams": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new past exam with file upload",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "past-exams"
                ],
                "summary": "Create a new past exam",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Year",
                        "name": "year",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Term (FALL, SPRING)",
                        "name": "term",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Department ID",
                        "name": "departmentId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course code",
                        "name": "courseCode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Instructor ID",
                        "name": "instructorId",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Exam file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.PastExamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/pastexams": {
            "get": {
                "description": "Retrieves a list of past exams with optional filtering and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "past-exams"
                ],
                "summary": "Get all past exams",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by faculty ID",
                        "name": "facultyId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by department ID",
                        "name": "departmentId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by course code",
                        "name": "courseCode",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by year",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by term (FALL, SPRING)",
                        "name": "term",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field (year, term, courseCode, title, departmentName, facultyName, instructorName, createdAt, updatedAt)",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (ASC, DESC)",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default: 10)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Past exams retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.PastExamListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_yigit_unisphere_internal_app_models_dto.APIResponse": {
            "description": "Generic response structure for all API endpoints",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response payload - specific to each endpoint"
                },
                "timestamp": {
                    "description": "Timestamp of when the response was generated",
                    "type": "string",
                    "example": "2025-04-23T12:01:05.123Z"
                }
            }
        },
...
        "github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse": {
            "description": "Generic response structure for all API endpoints",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response payload - specific to each endpoint"
                },
                "timestamp": {
                    "description": "Timestamp of when the response was generated",
                    "type": "string",
                    "example": "2025-04-23T12:01:05.123Z"
                }
            }
        },
...
        "github_com_yigit_unisphere_internal_app_models_dto.PastExamResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "courseCode": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "departmentId": {
                    "type": "integer"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.PastExamFileResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "instructorId": {
                    "type": "integer"
                },
                "term": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token (without Bearer prefix)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}